pyproject.toml
[project]
name = "aijobhunter"
version = "0.1.0"
description = "AI Job Hunter — automated LinkedIn job discovery, scoring, and application"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pymupdf>=1.24",
    "openai>=1.0",
    "playwright>=1.40",
]

[project.scripts]
hunt = "job_hunter.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.uv]
default-groups = ["dev"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_hunter"]
